// webapp/src/app/quality-alert/model/grouped-alerts.model.ts
/*
 * Copyright 2021 The PartChain Authors. All Rights Reserved.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import { Table } from 'src/app/shared/components/table/table';
import { QualityAlertChildren } from './quality-alert.model';

/**
 *
 *
 * @export
 * @interface GroupedAlerts
 */
export interface GroupedAlert {
  alertId: string;
  numberOfParts: number;
  icon: string;
  previousIcon: string;
  qualityType: string;
  previousQualityType: string;
  isActive: boolean;
  timestamp: string;
  children: QualityAlertChildren[];
  status: string;
  description: string;
  company?: string;
  type?: string;
  table?: Table;
  originCompany?: string;
  avatar?: { name: string; color: string };
  control?: string;
  relatedId?: string;
  isRecent?: boolean;
}